use Paralleljs and reintroduce workers to SiteMap
