+++ lib/components/SiteMap/FetchLocationUtils.d.ts
@@ -1,3 +1,3 @@
-export function parseLocationHierarchy(inHierarchy: any, parent?: any): {};
 export function fetchDomainHierarchy(domain: any): Promise<any>;
-export function fetchLocations(locations: any): Promise<any>;
+export function fetchSingleLocationREST(location: any): Promise<any>;
+export function fetchManyLocationsGraphQL(locations: any): Promise<any>;
