+++ lib/components/SiteMap/FetchLocationUtils.js
@@ -3,7 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.fetchLocations = exports.fetchDomainHierarchy = exports.parseLocationHierarchy = void 0;
+exports.fetchManyLocationsGraphQL = exports.fetchSingleLocationREST = exports.fetchDomainHierarchy = void 0;
 
 var _rxjs = require("rxjs");
 
@@ -13,43 +13,11 @@ var _NeonApi = _interopRequireDefault(require("../NeonApi/NeonApi"));
 
 var _NeonGraphQL = _interopRequireDefault(require("../NeonGraphQL/NeonGraphQL"));
 
-var _SiteMapWorkerSafeUtils = require("./SiteMapWorkerSafeUtils");
+var _parseDomainHierarchy = _interopRequireDefault(require("../../workers/parseDomainHierarchy"));
+
+var _parseLocationsArray = _interopRequireDefault(require("../../workers/parseLocationsArray"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-function _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }
-
-/**
- Recursive function to parse a deeply nest hierarchy object into a flat key/value object
- where keys are location names and values are objects containing only those location attributes
- the hierarchy affords us (type, description, and parent)
-*/
-var parseLocationHierarchy = function parseLocationHierarchy(inHierarchy) {
-  var parent = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : null;
-  var outHierarchy = {};
-  var name = inHierarchy.locationParentHierarchy ? null : inHierarchy.locationName;
-  var description = inHierarchy.locationDescription || null;
-  var type = inHierarchy.locationType || null;
-
-  if (description.includes('Not Used')) {
-    return outHierarchy;
-  }
-
-  if (name !== null) {
-    outHierarchy[name] = {
-      type: type,
-      description: description,
-      parent: parent
-    };
-  }
-
-  inHierarchy.locationChildHierarchy.forEach(function (subLocation) {
-    outHierarchy = _extends(_extends({}, outHierarchy), parseLocationHierarchy(subLocation, name));
-  });
-  return outHierarchy;
-};
-
-exports.parseLocationHierarchy = parseLocationHierarchy;
 
 var domainIsValid = function domainIsValid(domainString) {
   if (typeof domainString !== 'string') {
@@ -74,22 +42,13 @@ var domainIsValid = function domainIsValid(domainString) {
 var fetchDomainHierarchy = function fetchDomainHierarchy(domain) {
   if (!domainIsValid) {
     return Promise.reject(new Error('Domain is not valid'));
-  } // Execute Locations REST API hierarchy fetch and pipe results to processing function
+  } // Execute Locations REST API hierarchy fetch
 
 
   return new Promise(function (resolve, reject) {
     _NeonApi.default.getSiteLocationHierarchyObservable(domain).pipe((0, _operators.map)(function (response) {
       if (response && response.data) {
-        var data = {};
-        response.data.locationChildHierarchy.forEach(function (child) {
-          // At the top level we only care about sites and don't want the HQ test site
-          if (child.locationType !== 'SITE' || child.locationName === 'HQTW') {
-            return;
-          }
-
-          data[child.locationName] = parseLocationHierarchy(child);
-        });
-        resolve(data);
+        resolve(response.data);
         return (0, _rxjs.of)(true);
       }
 
@@ -99,12 +58,42 @@ var fetchDomainHierarchy = function fetchDomainHierarchy(domain) {
       reject(new Error(error.message));
       return (0, _rxjs.of)(false);
     })).subscribe();
+  }).then(function (data) {
+    return (0, _parseDomainHierarchy.default)(data);
   });
 };
 
 exports.fetchDomainHierarchy = fetchDomainHierarchy;
 
-var fetchLocations = function fetchLocations(locations) {
+var fetchSingleLocationREST = function fetchSingleLocationREST(location) {
+  if (typeof location !== 'string' || !location.length) {
+    return Promise.reject(new Error('Location is not valid; must be non-empty string'));
+  } // Execute REST query and pipe results to processing function
+
+
+  return new Promise(function (resolve, reject) {
+    _NeonApi.default.getLocationObservable(location).pipe((0, _operators.map)(function (response) {
+      if (response && response.data) {
+        resolve([response.data]);
+        return (0, _rxjs.of)(true);
+      }
+
+      reject(new Error('Malformed response'));
+      return (0, _rxjs.of)(false);
+    }), (0, _operators.catchError)(function (error) {
+      reject(new Error(error.message));
+      return (0, _rxjs.of)(false);
+    })).subscribe();
+  }).then(function (data) {
+    return (0, _parseLocationsArray.default)(data);
+  }).then(function (locationMap) {
+    return (locationMap || {})[location];
+  });
+};
+
+exports.fetchSingleLocationREST = fetchSingleLocationREST;
+
+var fetchManyLocationsGraphQL = function fetchManyLocationsGraphQL(locations) {
   // Extract locations list and validate
   if (!Array.isArray(locations) || !locations.length || !locations.every(function (loc) {
     return typeof loc === 'string';
@@ -121,24 +110,16 @@ var fetchLocations = function fetchLocations(locations) {
         return (0, _rxjs.of)(false);
       }
 
-      var data = {};
-      result.response.data.locations.forEach(function (rawLocationData) {
-        var locationName = rawLocationData.locationName;
-
-        if (!locationName) {
-          return;
-        }
-
-        data[locationName] = (0, _SiteMapWorkerSafeUtils.parseLocationData)(rawLocationData);
-      });
-      resolve(data);
+      resolve(result.response.data.locations);
       return (0, _rxjs.of)(true);
     }), // Error
     (0, _operators.catchError)(function (error) {
       reject(new Error(error.message));
       return (0, _rxjs.of)(false);
     })).subscribe();
+  }).then(function (data) {
+    return (0, _parseLocationsArray.default)(data);
   });
 };
 
-exports.fetchLocations = fetchLocations;
+exports.fetchManyLocationsGraphQL = fetchManyLocationsGraphQL;
