+++ lib/components/SiteMap/SiteMapContext.js
@@ -25,8 +25,6 @@ var _NeonContext = _interopRequireDefault(require("../NeonContext/NeonContext"))
 
 var _FetchLocationUtils = require("./FetchLocationUtils");
 
-var _SiteMapWorkerSafeUtils = require("./SiteMapWorkerSafeUtils");
-
 var _SiteMapUtils = require("./SiteMapUtils");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
@@ -417,8 +415,6 @@ var reducer = function reducer(state, action) {
       newState.featureDataFetches[dataSource][_featureKey][_siteCode][location] = status; // If the status is SUCCESS and the action has data, also commit the data
 
       if (status === _SiteMapUtils.FETCH_STATUS.SUCCESS && _data) {
-        var parsedData = (0, _SiteMapWorkerSafeUtils.parseLocationData)(_data);
-
         if (!newState.featureData[_featureType][_featureKey]) {
           newState.featureData[_featureType][_featureKey] = {};
         }
@@ -427,7 +423,7 @@ var reducer = function reducer(state, action) {
           newState.featureData[_featureType][_featureKey][_siteCode] = {};
         }
 
-        newState.featureData[_featureType][_featureKey][_siteCode][location] = _extends(_extends({}, parsedData), {}, {
+        newState.featureData[_featureType][_featureKey][_siteCode][location] = _extends(_extends({}, _data), {}, {
           siteCode: _siteCode,
           featureKey: _featureKey,
           name: location,
@@ -1067,7 +1063,7 @@ var Provider = function Provider(props) {
     dispatch({
       type: 'setFocusLocationFetchStarted'
     });
-    (0, _FetchLocationUtils.fetchLocations)([current]).then(function (response) {
+    (0, _FetchLocationUtils.fetchManyLocationsGraphQL)([current]).then(function (response) {
       dispatch({
         type: 'setFocusLocationFetchSucceeded',
         data: response[current]
@@ -1078,19 +1074,6 @@ var Provider = function Provider(props) {
         error: error
       });
     });
-    /* TODO: Restore worker support
-    const worker = new FetchLocationsWorker();
-    worker.addEventListener('message', (message) => {
-      const { status, data, error } = message.data;
-      if (status === 'success') {
-        dispatch({ type: 'setFocusLocationFetchSucceeded', data: data[current] });
-      } else {
-        dispatch({ type: 'setFocusLocationFetchFailed', error });
-      }
-      worker.terminate();
-    });
-    worker.postMessage([current]);
-    */
   }, [state.sites, state.focusLocation, state.focusLocation.fetch.status, state.neonContextHydrated, state.featureData]);
   /**
      Effect - trigger all data fetches and imports
@@ -1129,19 +1112,6 @@ var Provider = function Provider(props) {
           domainCode: domainCode
         });
       });
-      /*
-      const worker = new FetchLocationHierarchyWorker();
-      worker.addEventListener('message', (message) => {
-        const { status, data, error } = message.data;
-        if (status === 'success') {
-          dispatch({ type: 'setDomainLocationHierarchyFetchSucceeded', data, domainCode });
-        } else {
-          dispatch({ type: 'setDomainLocationHierarchyFetchFailed', error, domainCode });
-        }
-        worker.terminate();
-      });
-      worker.postMessage(domainCode);
-      */
     }); // ARCGIS_ASSETS_API Fetches
 
     var arcgisSource = _SiteMapUtils.FEATURE_DATA_SOURCES.ARCGIS_ASSETS_API;
@@ -1212,30 +1182,16 @@ var Provider = function Provider(props) {
             siteCode: siteCode,
             location: location
           });
-
-          _NeonApi.default.getLocationObservable(location).pipe((0, _operators.map)(function (response) {
-            if (response && response.data && response.data) {
-              dispatch({
-                type: 'setFeatureDataFetchSucceeded',
-                dataSource: restLocSource,
-                data: response.data,
-                featureKey: featureKey,
-                siteCode: siteCode,
-                location: location
-              });
-              return (0, _rxjs.of)(true);
-            }
-
+          (0, _FetchLocationUtils.fetchSingleLocationREST)(location).then(function (data) {
             dispatch({
-              type: 'setFeatureDataFetchFailed',
+              type: 'setFeatureDataFetchSucceeded',
               dataSource: restLocSource,
-              error: 'malformed response',
+              data: data,
               featureKey: featureKey,
               siteCode: siteCode,
               location: location
             });
-            return (0, _rxjs.of)(false);
-          }), (0, _operators.catchError)(function (error) {
+          }).catch(function (error) {
             dispatch({
               type: 'setFeatureDataFetchFailed',
               dataSource: restLocSource,
@@ -1244,8 +1200,7 @@ var Provider = function Provider(props) {
               siteCode: siteCode,
               location: location
             });
-            return (0, _rxjs.of)(false);
-          })).subscribe();
+          });
         });
       });
     }); // GRAPHQL_LOCATIONS_API Fetches
@@ -1264,7 +1219,7 @@ var Provider = function Provider(props) {
             siteCode: siteCode,
             fetchId: fetchId
           });
-          (0, _FetchLocationUtils.fetchLocations)(gqlLocFetches[minZoom][siteCode].fetches[fetchId].locations).then(function (response) {
+          (0, _FetchLocationUtils.fetchManyLocationsGraphQL)(gqlLocFetches[minZoom][siteCode].fetches[fetchId].locations).then(function (response) {
             dispatch({
               type: 'setFeatureDataFetchSucceeded',
               dataSource: gqlLocSource,
@@ -1283,33 +1238,6 @@ var Provider = function Provider(props) {
               error: error
             });
           });
-          /*
-          const worker = new FetchLocationsWorker();
-          worker.addEventListener('message', (message) => {
-            const { status, data, error } = message.data;
-            if (status === 'success') {
-              dispatch({
-                type: 'setFeatureDataFetchSucceeded',
-                dataSource: gqlLocSource,
-                minZoom,
-                siteCode,
-                fetchId,
-                data,
-              });
-            } else {
-              dispatch({
-                type: 'setFeatureDataFetchFailed',
-                dataSource: gqlLocSource,
-                minZoom,
-                siteCode,
-                fetchId,
-                error,
-              });
-            }
-            worker.terminate();
-          });
-          worker.postMessage(gqlLocFetches[minZoom][siteCode].fetches[fetchId].locations);
-          */
         });
       });
     });
