--- lib/components/SiteMap/SiteMapWorkerSafeUtils.d.ts
@@ -1,8 +1,0 @@
-export function parseLocationProperties(inProps?: any[], whiteList?: string[]): {};
-export function parseLocationData(data?: {}): {
-    name: any;
-    type: any;
-    description: any;
-    siteCode: any;
-    parent: any;
-};
