+++ src/lib_components/components/SiteMap/SiteMapContainer.jsx
@@ -145,7 +145,7 @@ const SiteMapContainer = (props) => {
   const [neonContextState] = NeonContext.useNeonContextState();
 
   const [state, dispatch] = SiteMapContext.useSiteMapContext();
-  console.log('CONTAINER STATE:', state);
+  // console.log('CONTAINER STATE:', state);
   const isLoading = state.overallFetch.expected !== state.overallFetch.completed;
 
   const { aspectRatio, view: { current: view } } = state;
