+++ src/lib_components/components/SiteMap/StyleGuide.jsx
@@ -2,7 +2,6 @@
 
 import React, { useState } from 'react';
 
-import Parallel from 'paralleljs';
 import cloneDeep from 'lodash/cloneDeep';
 
 import { makeStyles } from '@material-ui/core/styles';
@@ -227,22 +226,6 @@ export default function StyleGuide() {
       #portal-feedback Slack channel
     </Link>
   );
-
-  /**
-    WORKERS
-  */
-  const [run, setRun] = useState(false);
-  if (!run) {
-    setRun(true);
-    const p = new Parallel(4);
-    console.log(p);
-    p.spawn((dataIn) => {
-      const j = dataIn * 2;
-      return j;
-    }).then((dataOut) => {
-      console.log(dataOut);
-    });
-  }
 
   return (
     <React.Fragment>
